renamed component + small refactor
